[update] move deploy-commands into the module directory for better code structure
