+++ modules/deploy-commands.js
@@ -1,18 +1,18 @@
 const fs = require("fs");
 const { REST } = require("@discordjs/rest");
 const { Routes } = require("discord-api-types/v9");
-const { token } = require("./config.json");
+const { token } = require("../config.json");
 const rest = new REST({ version: "9" }).setToken(token);
 
 const commands = [];
-const commandFolders = fs.readdirSync("./commands");
+const commandFolders = fs.readdirSync("../commands");
 
 for (const folder of commandFolders) {
   const commandFiles = fs
-    .readdirSync(`./commands/${folder}`)
+    .readdirSync(`../commands/${folder}`)
     .filter((file) => file.endsWith(".js"));
   for (const file of commandFiles) {
-    const command = require(`./commands/${folder}/${file}`);
+    const command = require(`../commands/${folder}/${file}`);
     commands.push(command.data.toJSON());
   }
 }
